requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
discord.py==2.7.1
dotenv==0.9.9
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
python-dotenv==1.2.2
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
